:white_check_mark: [#130] update tests
	modified:   src/openproduct/producttypen/models/validators.py
	modified:   src/openproduct/producttypen/tests/api/test_producttype.py
	modified:   src/openproduct/producttypen/tests/test_producttype.py
	modified:   src/openproduct/producttypen/tests/test_producttype_admin.py
	deleted:    src/openproduct/producttypen/tests/test_validators.py
	modified:   src/producten-openapi.yaml
	modified:   src/producttypen-openapi.yaml

	modified:   src/openproduct/producttypen/models/validators.py
	modified:   src/openproduct/producttypen/tests/api/test_producttype.py
	modified:   src/openproduct/producttypen/tests/test_producttype.py#
	modified:   src/openproduct/producttypen/tests/test_producttype_admin.py
	deleted:    src/openproduct/producttypen/tests/test_validators.py
	modified:   src/producten-openapi.yaml
	modified:   src/producttypen-openapi.yaml

	modified:   src/openproduct/producttypen/models/validators.py
	modified:   src/openproduct/producttypen/serializers/producttype.py
	modified:   src/openproduct/producttypen/serializers/thema.py
	modified:   src/openproduct/producttypen/tests/api/test_producttype.py
	modified:   src/openproduct/producttypen/tests/test_producttype.py
	modified:   src/openproduct/producttypen/tests/test_producttype_admin.py
	deleted:    src/openproduct/producttypen/tests/test_validators.py
	modified:   src/producten-openapi.yaml
	modified:   src/producttypen-openapi.yaml

	modified:   src/openproduct/producttypen/models/validators.py
	modified:   src/openproduct/producttypen/serializers/producttype.py
	modified:   src/openproduct/producttypen/serializers/thema.py
	modified:   src/openproduct/producttypen/tests/api/test_producttype.py
	modified:   src/openproduct/producttypen/tests/test_producttype.py
	modified:   src/openproduct/producttypen/tests/test_producttype_admin.py
	deleted:    src/openproduct/producttypen/tests/test_validators.py
	modified:   src/producten-openapi.yaml
	modified:   src/producttypen-openapi.yaml

	modified:   src/openproduct/producten/serializers/product.py
	modified:   src/openproduct/producttypen/models/validators.py
	modified:   src/openproduct/producttypen/serializers/producttype.py
	modified:   src/openproduct/producttypen/serializers/thema.py
	modified:   src/openproduct/producttypen/tests/api/test_producttype.py
	modified:   src/openproduct/producttypen/tests/test_producttype.py
	modified:   src/openproduct/producttypen/tests/test_producttype_admin.py
	deleted:    src/openproduct/producttypen/tests/test_validators.py
	modified:   src/producten-openapi.yaml
	modified:   src/producttypen-openapi.yaml
