+++ src/producten-openapi.yaml
@@ -364,7 +364,7 @@ paths:
                       update_datum: '2019-08-24T14:15:22Z'
                       producttype:
                         uuid: 497f6eca-6276-4993-bfeb-53cbbbba6f08
-                        code: 129380-c21231
+                        code: 129380-A21231
                         keywords:
                         - auto
                         uniforme_product_naam: parkeervergunning
@@ -442,7 +442,7 @@ paths:
                     update_datum: '2019-08-24T14:15:22Z'
                     producttype:
                       uuid: 497f6eca-6276-4993-bfeb-53cbbbba6f08
-                      code: 129380-c21231
+                      code: 129380-A21231
                       keywords:
                       - auto
                       uniforme_product_naam: parkeervergunning
@@ -522,7 +522,7 @@ paths:
                     update_datum: '2019-08-24T14:15:22Z'
                     producttype:
                       uuid: 497f6eca-6276-4993-bfeb-53cbbbba6f08
-                      code: 129380-c21231
+                      code: 129380-A21231
                       keywords:
                       - auto
                       uniforme_product_naam: parkeervergunning
@@ -607,7 +607,7 @@ paths:
                     update_datum: '2019-08-24T14:15:22Z'
                     producttype:
                       uuid: 497f6eca-6276-4993-bfeb-53cbbbba6f08
-                      code: 129380-c21231
+                      code: 129380-A21231
                       keywords:
                       - auto
                       uniforme_product_naam: parkeervergunning
@@ -691,7 +691,7 @@ paths:
                     update_datum: '2019-08-24T14:15:22Z'
                     producttype:
                       uuid: 497f6eca-6276-4993-bfeb-53cbbbba6f08
-                      code: 129380-c21231
+                      code: 129380-A21231
                       keywords:
                       - auto
                       uniforme_product_naam: parkeervergunning
@@ -890,6 +890,7 @@ components:
         code:
           type: string
           description: code van het producttype.
+          pattern: ^[A-Z0-9-]+$
           maxLength: 255
         keywords:
           type: array
@@ -931,6 +932,7 @@ components:
           type: string
           minLength: 1
           description: code van het producttype.
+          pattern: ^[A-Z0-9-]+$
           maxLength: 255
         keywords:
           type: array
